Try to reduce code duplication
