+++ src/index.ts
@@ -1,5 +1,4 @@
 import type {
-    ComposerTranslation,
     I18n,
 } from 'vue-i18n'
 import type { ErrorMapCtx, ZodErrorMap, ZodIssueOptionalMessage } from 'zod/v3'
@@ -11,32 +10,15 @@ import {
     ZodIssueCode,
     ZodParsedType,
 } from 'zod/v3'
-import { retrieveCount } from './utils'
+import { translateLabelFactory } from './utils'
 
 const zDate = z.string().regex(/(\d{4})-\d{2}-(\d{2})/)
 
 function makeZodI18nMap(i18n: I18n, key = 'errors'): ZodErrorMap {
-    const t = i18n.global.t as ComposerTranslation
-    const te = i18n.global.te
     const d = i18n.global.d
     const n = i18n.global.n
 
-    const translateLabel = (label: string, { named = {}, prefix, count }: TranslateLabelOptions = {}) => {
-        const hasCount = count ?? retrieveCount(named)
-        const completeLabel = `${prefix ? `${prefix}.` : ''}${label}`
-        const messageKey = [
-            `${key}${completeLabel}WithPath`,
-            `${key}.${completeLabel}`,
-            completeLabel,
-        ].find(k => te(k))
-
-        if (!messageKey)
-            return label
-
-        return hasCount !== undefined
-            ? t(messageKey, hasCount, { named })
-            : t(messageKey, named)
-    }
+    const translateLabel = translateLabelFactory(i18n, key)
 
     return (issue: ZodIssueOptionalMessage, ctx: ErrorMapCtx): { message: string } => {
         let message = defaultErrorMap(issue, ctx).message
